Move supervisord config into a separate recipe
Rename install/supervisord.rb to supervisord_install to avoid overlapping names.
Split ChefSpec tests.
