+++ cookbooks/aws-parallelcluster-platform/kitchen.platform-config.yml
@@ -162,12 +162,14 @@ suites:
         - /tag:config_networking/
   - name: sudo
     run_list:
-      - recipe[aws-parallelcluster-tests::setup]
       - recipe[aws-parallelcluster-platform::sudo_config]
     verifier:
       controls:
         - /tag:config_sudo/
-    attributes:
-      dependencies:
-        - recipe:aws-parallelcluster-platform::directories
-        - recipe:aws-parallelcluster-platform::sudo_install
+  - name: supervisord
+    run_list:
+      - recipe[aws-parallelcluster-platform::supervisord_config]
+      - recipe[aws-parallelcluster-platform::finalize]  # supervisord service start
+    verifier:
+      controls:
+        - /tag:config_supervisord/
