+++ cookbooks/aws-parallelcluster-platform/recipes/config.rb
@@ -12,11 +12,11 @@
 # limitations under the License.
 
 include_recipe 'aws-parallelcluster-platform::openssh'
-include_recipe "aws-parallelcluster-platform::sudo_config"
+include_recipe 'aws-parallelcluster-platform::sudo_config'
 include_recipe 'aws-parallelcluster-platform::cluster_user'
 include_recipe 'aws-parallelcluster-platform::networking'
-include_recipe "aws-parallelcluster-platform::nvidia_config"
-sticky_bits "setup sticky bits"
+include_recipe 'aws-parallelcluster-platform::nvidia_config'
+sticky_bits 'setup sticky bits'
 chrony 'enable Amazon Time Sync' do
   action :enable
 end
@@ -24,4 +24,6 @@ include_recipe 'aws-parallelcluster-platform::dcv'
 intel_hpc 'Configure Intel HPC' do
   action :configure
 end
+# Supervisord configuration must be executed after DCV because dcv external authenticator is part of it
+include_recipe 'aws-parallelcluster-platform::supervisord_config'
 fetch_config 'Fetch and load cluster configs'
