+++ cookbooks/aws-parallelcluster-platform/recipes/install/supervisord_install.rb
